Optimize to use uint256, clean up contracts (#71)
* Fix to uint256

* Fix compile warnings

* Change uint104 to uint256

* Remove redundant code

* Fix test

* Remove casting

* Avoid creating new var

* Fix deposit error

* Nit

* Lint

* Fix gas test

* Fix comments

+++ contracts/vaults/YearnVaults/RibbonThetaYearnVault.sol
@@ -135,7 +135,7 @@ contract RibbonThetaYearnVault is RibbonVault, OptionsThetaYearnVaultStorage {
      * @notice Sets the new discount on premiums for options we are selling
      * @param newPremiumDiscount is the premium discount
      */
-    function setPremiumDiscount(uint16 newPremiumDiscount) external onlyOwner {
+    function setPremiumDiscount(uint256 newPremiumDiscount) external onlyOwner {
         require(
             newPremiumDiscount > 0 && newPremiumDiscount < 1000,
             "Invalid discount"
@@ -265,9 +265,9 @@ contract RibbonThetaYearnVault is RibbonVault, OptionsThetaYearnVaultStorage {
     function _closeShort(address oldOption) private {
         optionState.currentOption = address(0);
 
-        uint104 lockedAmount = vaultState.lockedAmount;
+        uint256 lockedAmount = vaultState.lockedAmount;
         vaultState.lastLockedAmount = lockedAmount > 0
-            ? lockedAmount
+            ? uint104(lockedAmount)
             : vaultState.lastLockedAmount;
         vaultState.lockedAmount = 0;
